+++ nncf/__init__.py
@@ -16,7 +16,7 @@ from nncf.version import __version__
 
 from nncf.config import NNCFConfig
 from nncf.data import Dataset
-from nncf.parameters import IgnoredScope
+from nncf.scopes import IgnoredScope
 from nncf.parameters import ModelType
 from nncf.parameters import TargetDevice
 from nncf.quantization import QuantizationPreset
