+++ tests/torch/ptq/test_ptq_params.py
@@ -15,7 +15,7 @@
 import pytest
 from torch import nn
 
-from nncf.parameters import IgnoredScope
+from nncf.scopes import IgnoredScope
 from nncf.parameters import TargetDevice
 from nncf.common.graph.patterns import GraphPattern
 from nncf.quantization.algorithms.post_training.algorithm import PostTrainingQuantization
